+++ CBLApp_Backend/src/main/java/com/cbl/backend/model/Product.java
@@ -27,21 +27,21 @@ public class Product{
 	private String productName;
 	@NotEmpty
 	@Column
-	private int unitBuyingPrice;
+	private float unitBuyingPrice;
 	@NotEmpty
 	@Column
-	private int unitSellingPrice;
+	private float unitSellingPrice;
 	@NotEmpty
 	@Column
 	private LocalDate date;
 	@Column
 	private int pieces;
 	@Column
-	private int buyingPrice=unitBuyingPrice*pieces;
+	private float buyingPrice=unitBuyingPrice*pieces;
 	@Column
-	private int sellingPrice=unitSellingPrice*pieces;
+	private float sellingPrice=unitSellingPrice*pieces;
 	@Column
-	private int profit=sellingPrice-buyingPrice;
+	private float profit=sellingPrice-buyingPrice;
 	
 	@ManyToOne
 	@JoinColumn(name = "reportID")
@@ -60,19 +60,18 @@ public class Product{
 	public void setProductName(String productName) {
 		this.productName = productName;
 	}
-	public int getUnitBuyingPrice() {
+	public float getUnitBuyingPrice() {
 		return unitBuyingPrice;
 	}
-	public void setUnitBuyingPrice(int unitBuyingPrice) {
+	public void setUnitBuyingPrice(float unitBuyingPrice) {
 		this.unitBuyingPrice = unitBuyingPrice;
 	}
-	public int getUnitSellingPrice() {
+	public float getUnitSellingPrice() {
 		return unitSellingPrice;
 	}
-	public void setUnitSellingPrice(int unitSellingPrice) {
+	public void setUnitSellingPrice(float unitSellingPrice) {
 		this.unitSellingPrice = unitSellingPrice;
 	}
-	
 	public LocalDate getDate() {
 		return date;
 	}
@@ -85,24 +84,27 @@ public class Product{
 	public void setPieces(int pieces) {
 		this.pieces = pieces;
 	}
-	public int getBuyingPrice() {
+	public float getBuyingPrice() {
 		return buyingPrice;
 	}
-	public void setBuyingPrice(int buyingPrice) {
+	public void setBuyingPrice(float buyingPrice) {
 		this.buyingPrice = buyingPrice;
 	}
-	public int getSellingPrice() {
+	public float getSellingPrice() {
 		return sellingPrice;
 	}
-	public void setSellingPrice(int sellingPrice) {
+	public void setSellingPrice(float sellingPrice) {
 		this.sellingPrice = sellingPrice;
 	}
-	public int getProfit() {
+	public float getProfit() {
 		return profit;
 	}
-	public void setProfit(int profit) {
+	public void setProfit(float profit) {
 		this.profit = profit;
 	}
+	
+	
+	
 	
 	
 }
